requirements.txt
matplotlib
numpy
scipy
Pillow
sounddevice 
soundfile